refactor(egopulse): split llm.rs into submodules
